Get ready to implement plot pixel operation

+++ src/config.rs
@@ -1,14 +1,14 @@
 pub const PIXEL_BYTES: usize = 4;
 
 bitflags! {
-    pub struct MGLBit: u32 {
+    pub struct MGLBit: usize {
         const COLOR     = 0b00000001;
         const DEPTH     = 0b00000010;
     }
 }
 
 bitflags! {
-    pub struct MGLColor: u32 {
+    pub struct MGLColor: usize {
         const RED      = 0xff0000;
     }
 }
